x32 to x64

+++ core/NrtlType.h
@@ -11,7 +11,7 @@ class MeshModel;
 
 enum {NONE=0};
 
-typedef uint32_t DataId;
+typedef uint64_t DataId;
 typedef std::vector<DataId> ResourceList;
 typedef QVector<QVector<int>> PolygonMatrix;
 typedef QVector<QVector3D> MeshBorder;
